Add pause menu system to Engine worlds.
Comments by Dan Ristic on 1/4/2012:
This commit was merged over from the Shooter-Demo project's src folder.
The reason for this was that the pause menu was developed and put into
that project fully before being integrated into this project. This was a
full copy overwrite to the engine's folder structure.